rename and refactoring project structure wip
