+++ ZipUtility/LogApi/LogApi.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FileLogger\SRF.FileLogger.csproj" />
+    <ProjectReference Include="..\Logging\SRF.FileLogging.csproj" />
   </ItemGroup>
