+++ ZipUtility/LogApi/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-using SRF.FileLogger;
+using SRF.FileLogging;
 
 namespace LogApi
 {
