+++ ZipUtility/ZipUtility.sln
@@ -17,9 +17,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "zipper", "zipper\zipper.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "moncon", "moncon\moncon.csproj", "{8200B3E6-9CFC-4B1A-A1A2-8FCFA1BB49BC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SRF.FileLogger", "FileLogger\SRF.FileLogger.csproj", "{FC639569-5FC9-4977-BB9C-E8331C18640D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LogApi", "LogApi\LogApi.csproj", "{DC9672B0-E065-467E-B5E3-CCE91BB63893}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogApi", "LogApi\LogApi.csproj", "{DC9672B0-E065-467E-B5E3-CCE91BB63893}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SRF.FileLogging", "Logging\SRF.FileLogging.csproj", "{9EE8D6F0-FEB9-48DA-A52E-33983B00DA84}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,14 +51,14 @@ Global
 		{8200B3E6-9CFC-4B1A-A1A2-8FCFA1BB49BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8200B3E6-9CFC-4B1A-A1A2-8FCFA1BB49BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8200B3E6-9CFC-4B1A-A1A2-8FCFA1BB49BC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FC639569-5FC9-4977-BB9C-E8331C18640D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FC639569-5FC9-4977-BB9C-E8331C18640D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FC639569-5FC9-4977-BB9C-E8331C18640D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FC639569-5FC9-4977-BB9C-E8331C18640D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DC9672B0-E065-467E-B5E3-CCE91BB63893}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DC9672B0-E065-467E-B5E3-CCE91BB63893}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DC9672B0-E065-467E-B5E3-CCE91BB63893}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DC9672B0-E065-467E-B5E3-CCE91BB63893}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9EE8D6F0-FEB9-48DA-A52E-33983B00DA84}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9EE8D6F0-FEB9-48DA-A52E-33983B00DA84}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9EE8D6F0-FEB9-48DA-A52E-33983B00DA84}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9EE8D6F0-FEB9-48DA-A52E-33983B00DA84}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
